Simplify message channel synchronization construct
This whole sender wait group unnecessarily complicates processing
messages in the face of interaction with closing the connection and
cleaning up. Subtle bugs and races are also possible when context
switching. The solution is to just never close the message channel
and to let the Go runtime handle the cleanup after the shutdown.